+++ src/index.ts
@@ -1,17 +1,14 @@
 import 'dotenv/config';
+import './config/passport';
 import express from 'express';
 import passport from 'passport';
 import cors from 'cors';
-
 import session from './config/session';
-import './config/passport';
-
 import testAndInitializeDatabase from './models';
+import isAuthenticated from './middlewares/isAuthenticated';
 import router from './router/app-router';
 import authRouter from './router/authRoutes';
-import './config/passport';
 import errorHandler from './middlewares/errorHandler';
-
 
 const app = express();
 const PORT = process.env.PORT || 3000;
@@ -28,6 +25,8 @@ app.use(cors({
 app.use(session);
 app.use(passport.initialize());
 app.use(passport.session());
+
+app.use(isAuthenticated);
 
 app.use('/auth', authRouter);
 app.use('/api/v1', router);
