+++ src/middlewares/isAuthenticated.ts
@@ -1,9 +1,11 @@
 import { Request, Response, NextFunction } from 'express';
 
-export const isAuth = (req: Request, res: Response, next: NextFunction): void => {
+const isAuthenticated = (req: Request, res: Response, next: NextFunction): void => {
   if (req.isAuthenticated()) {
     return next();
   }
 
   res.status(401).json({ message: 'Unauthorized. Please log in.' });
 };
+
+export default isAuthenticated;
